Update releasenodes.md
Adding 1.6.2 features pointed out by Jeff in BUG-72292

Tried to clarify Fixes & Changes but for most part I don't know which Jiras the fixes refer to. I found a few jiras but they were cryptic too. So - to be honest - I don't know what most of the items under Fixes & Changes mean... Maybe someone who knows could clarify? Or if Jeff is OK with that, we can leave them as they are.